+++ forge20/src/main/java/com/vecoo/extraquests/storage/QuestTimer.java
@@ -2,10 +2,31 @@ package com.vecoo.extraquests.storage;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.util.Objects;
 import java.util.UUID;
 
 public record QuestTimer(@NotNull UUID playerUUID, @NotNull String questID, long endTime) {
     public QuestTimer(@NotNull UUID playerUUID, @NotNull String questID, int secondsTimer) {
         this(playerUUID, questID, System.currentTimeMillis() + (secondsTimer * 1000L));
+    }
+
+    @Override
+    public boolean equals(Object object) {
+        if (this == object) {
+            return true;
+        }
+
+        if (object == null || getClass() != object.getClass()) {
+            return false;
+        }
+
+        QuestTimer timerStorage = (QuestTimer) object;
+
+        return Objects.equals(this.playerUUID, timerStorage.playerUUID) && Objects.equals(this.questID, timerStorage.questID);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(this.playerUUID, this.questID);
     }
 }