+++ changelog.md
@@ -1,3 +1,1 @@
-* Change name methods.
-* Update 1.20.1 timer storage.
-* Update depends.
+* Fix storage (Forge 1.20.1).